requirements.txt
streamlit>=1.35.0
groq>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
fpdf2>=2.7.9
audioop-lts>=0.2.1